+++ src/modules/tools/print/store/actions/actionsPrintInitialization.js
@@ -118,9 +118,11 @@ export default {
      * @returns {void}
      */
     chooseCurrentLayout: function ({state, commit}, layouts) {
-        const currentLayout = layouts.filter(layout => layout.name === state.currentLayoutName);
+        const configuredLayout = layouts.find(layout => layout.name === state.currentLayoutName),
+            layoutToUse = configuredLayout || layouts[0];
 
-        commit("setCurrentLayout", currentLayout.length === 1 ? currentLayout[0] : layouts[0]);
+        commit("setCurrentLayout", layoutToUse);
+        commit("setCurrentLayoutName", layoutToUse.name);
     },
 
     /**