+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Fixed
 - Issue #733, #680, #683: transparency setting has no visible effect in treeType custom
+- Issue #700: Print tool now updates layout name on initialization.
 
 ---
 ## v2.19.0 - 2022-03-02